feat: refine context menu ordering and remove unnecessary system items
